refactor: Unify the implementation of basic string manipulation methods
